+++ src/test/groovy/edu/ie3/util/quantities/QuantityUtilTest.groovy
@@ -5,8 +5,9 @@
  */
 package edu.ie3.util.quantities
 
+import edu.ie3.util.quantities.interfaces.EnergyDensity
+
 import static edu.ie3.util.quantities.PowerSystemUnits.*
-import edu.ie3.util.quantities.interfaces.Irradiation
 import spock.lang.Unroll
 import tech.units.indriya.ComparableQuantity
 import tech.units.indriya.quantity.Quantities
@@ -36,7 +37,7 @@ class QuantityUtilTest extends Specification {
 		Quantities.getQuantity(10d, KILOWATT)                     || Quantities.getQuantity(10d, KILOWATT) as ComparableQuantity<Power>
 		Quantities.getQuantity(15d, DEGREE_GEOM)                  || Quantities.getQuantity(15d, DEGREE_GEOM) as ComparableQuantity<Angle>
 		Quantities.getQuantity(20d, METRE_PER_SECOND)             || Quantities.getQuantity(20d, METRE_PER_SECOND) as ComparableQuantity<Speed>
-		Quantities.getQuantity(25d, KILOWATTHOUR_PER_SQUAREMETRE) || Quantities.getQuantity(25d, KILOWATTHOUR_PER_SQUAREMETRE) as ComparableQuantity<Irradiation>
+		Quantities.getQuantity(25d, KILOWATTHOUR_PER_SQUAREMETRE) || Quantities.getQuantity(25d, KILOWATTHOUR_PER_SQUAREMETRE) as ComparableQuantity<EnergyDensity>
 		Quantities.getQuantity(30d, METRE)                        || Quantities.getQuantity(30d, METRE) as ComparableQuantity<Length>
 	}